+++ util/util.go
@@ -2,7 +2,6 @@
 package util
 
 import (
-	"gitee.com/conero/uymas/str"
 	"math"
 	"reflect"
 	"time"
@@ -100,54 +99,4 @@ func ValueNull(value interface{}) bool {
 	}
 	v := reflect.ValueOf(value)
 	return v.IsZero()
-}
-
-// StructToMap convert Struct field to by Map, support the Ptr
-func StructToMap(value interface{}) map[string]interface{} {
-	rv := reflect.ValueOf(value)
-	var rt reflect.Type
-	if rv.Kind() == reflect.Ptr {
-		rv = rv.Elem()
-		rt = rv.Type()
-	}
-	if rv.Kind() == reflect.Struct {
-		if rt == nil {
-			rt = reflect.TypeOf(value)
-		}
-		vMap := map[string]interface{}{}
-		for i := 0; i < rv.NumField(); i++ {
-			field := rv.Field(i)
-			if field.Kind() != reflect.Func && field.CanInterface() {
-				name := rt.Field(i).Name
-				vMap[name] = field.Interface()
-			}
-		}
-		return vMap
-	}
-	return nil
-}
-
-// StructToMapLStyle convert Struct field to by Map and key is Lower style.
-func StructToMapLStyle(value interface{}) map[string]interface{} {
-	rv := reflect.ValueOf(value)
-	var rt reflect.Type
-	if rv.Kind() == reflect.Ptr {
-		rv = rv.Elem()
-		rt = rv.Type()
-	}
-	if rv.Kind() == reflect.Struct {
-		if rt == nil {
-			rt = reflect.TypeOf(value)
-		}
-		vMap := map[string]interface{}{}
-		for i := 0; i < rv.NumField(); i++ {
-			field := rv.Field(i)
-			if field.Kind() != reflect.Func {
-				name := rt.Field(i).Name
-				vMap[str.LowerStyle(name)] = field.Interface()
-			}
-		}
-		return vMap
-	}
-	return nil
 }